V0.1.4 - CreateLaps action created in HomeController

+++ Models/CourseViewModel.cs
@@ -5,7 +5,7 @@ namespace RagnarEstimator.Models
 {
   public class CourseViewModel
   {
-    public decimal newCourseDistance {get; set;}
+    public double newCourseDistance {get; set;}
     public int newCourseElevation {get; set;}
     public int newCourseDifficulty {get; set;}
     public int newCourseSequence {get; set;}
